@@ -1,6 +1,6 @@
 module github.com/sidkurella/gunion
 
-go 1.19
+go 1.24.0
 
 require (
 	github.com/spf13/cobra v1.6.1
@@ -10,13 +10,13 @@ require (
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/mod v0.7.0 // indirect
-	golang.org/x/sys v0.4.0 // indirect
+	golang.org/x/mod v0.33.0 // indirect
+	golang.org/x/sync v0.19.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.0.1 // indirect
 	github.com/stretchr/testify v1.8.1
-	golang.org/x/tools v0.5.0
+	golang.org/x/tools v0.42.0
 )
